extract and implement browseSalesQuery
