+++ frontend/src/BrowseSales/BrowseSales.tsx
@@ -1,23 +1,17 @@
 import React, { FC, useState } from 'react';
 import SearchBar from '../common/SearchBar';
 import SalesList from './SalesList';
-import axios from 'axios';
-import makeAuthHeader from '../utils/makeAuthHeader';
-import { GET_SALES_BY_TITLE } from '../utils/api_resources';
 import { Header, Divider } from 'semantic-ui-react';
-import { Sale } from './SalesList';
+import browseSalesQuery, { Sale } from './browseSalesQuery';
 
 const BrowseSales: FC = () => {
     const [salesList, setSalesList] = useState<Sale[]>([]);
     const [cardName, setCardName] = useState<string>('');
 
     const handleSearchSelect = async (cardName: string) => {
-        const { data }: { data: Sale[] } = await axios.get(GET_SALES_BY_TITLE, {
-            params: { cardName: cardName },
-            headers: makeAuthHeader(),
-        });
+        const sales = await browseSalesQuery({ cardName });
 
-        setSalesList(data);
+        setSalesList(sales);
         setCardName(cardName);
     };
